Prepare for route
