+++ app/src/main/java/io/nekohasekai/sagernet/database/RuleEntity.kt
@@ -50,6 +50,9 @@ data class RuleEntity(
         @Query("SELECT * FROM rules ORDER BY userOrder")
         fun allRules(): List<RuleEntity>
 
+        @Query("SELECT * FROM rules WHERE enabled = :enabled ORDER BY userOrder")
+        fun enabledRules(enabled: Boolean = true): List<RuleEntity>
+
         @Query("SELECT MAX(userOrder) + 1 FROM rules")
         fun nextOrder(): Long?
